+++ src/Rpc/CoreRpcServerCommandsDefinitions.h
@@ -48,7 +48,7 @@ struct COMMAND_RPC_GET_HEIGHT {
   typedef EMPTY_STRUCT request;
 
   struct response {
-    uint64_t height;
+    uint32_t height;
     std::string status;
 
     void serialize(ISerializer &s) {
@@ -70,8 +70,8 @@ struct COMMAND_RPC_GET_BLOCKS_FAST {
 
   struct response {
     std::vector<block_complete_entry> blocks;
-    uint64_t start_height;
-    uint64_t current_height;
+    uint32_t start_height;
+    uint32_t current_height;
     std::string status;
 
     void serialize(ISerializer &s) {
@@ -272,7 +272,7 @@ struct COMMAND_RPC_GET_INFO {
   struct response {
     std::string status;
     std::string version;
-    uint64_t height;
+    uint32_t height;
     std::string top_block_hash;
     uint64_t difficulty;
     uint64_t min_tx_fee;
@@ -443,8 +443,8 @@ struct block_header_response {
   std::string prev_hash;
   uint32_t nonce;
   bool orphan_status;
-  uint64_t height;
-  uint64_t depth;
+  uint32_t height;
+  uint32_t depth;
   std::string hash;
   difficulty_type difficulty;
   uint64_t reward;
@@ -580,8 +580,8 @@ struct f_block_details_response {
   std::string prev_hash;
   uint32_t nonce;
   bool orphan_status;
-  uint64_t height;
-  uint64_t depth;
+  uint32_t height;
+  uint32_t depth;
   std::string hash;
   difficulty_type difficulty;
   difficulty_type cumulativeDifficulty;
@@ -642,7 +642,7 @@ struct COMMAND_RPC_GET_BLOCK_HEADER_BY_HASH {
 
 struct COMMAND_RPC_GET_BLOCK_HEADER_BY_HEIGHT {
   struct request {
-    uint64_t height;
+    uint32_t height;
 
     void serialize(ISerializer &s) {
       KV_MEMBER(height)
@@ -654,7 +654,7 @@ struct COMMAND_RPC_GET_BLOCK_HEADER_BY_HEIGHT {
 
 struct F_COMMAND_RPC_GET_BLOCKS_LIST {
   struct request {
-    uint64_t height;
+    uint32_t height;
 
     void serialize(ISerializer &s) {
       KV_MEMBER(height)
@@ -778,8 +778,8 @@ struct COMMAND_RPC_QUERY_BLOCKS {
 
   struct response {
     std::string status;
-    uint64_t start_height;
-    uint64_t current_height;
+    uint32_t start_height;
+    uint32_t current_height;
     uint64_t full_offset;
     std::vector<BlockFullInfo> items;
 
@@ -806,8 +806,8 @@ struct COMMAND_RPC_QUERY_BLOCKS_LITE {
 
   struct response {
     std::string status;
-    uint64_t startHeight;
-    uint64_t currentHeight;
+    uint32_t startHeight;
+    uint32_t currentHeight;
     uint64_t fullOffset;
     std::vector<BlockShortInfo> items;
